refactor: separate TS types for each endpoint to avoid confusion
