+++ dws/src/utils/api/types/index.ts
@@ -1,0 +1,7 @@
+export * from "./campaign.type";
+export * from "./userData.type";
+
+export interface FailedResponse {
+  code: string;
+  message: string;
+}
